faster synth fixed tables (keyrange editing)
Former-commit-id: 9e5a977c9296d6daab1ada43a58058926b85390d [formerly 4016478ca756a037c5b9512437ca4029bc28ff67]
Former-commit-id: 3d1ff9acd26cc4c530500ee702db4a85756fc768
Former-commit-id: c8e9b96d48fce817eb7f292a52e9dc2b3f46d5f8 [formerly 5305bff3394b9ab1c1643e9c5099f031910a6fad]
Former-commit-id: 2862c40e32b7b042105e0e81c11d22f5f85bf3c9

+++ trunk/synthetiseur/elements/oscsinus.cpp
@@ -33,7 +33,7 @@ OscSinus::OscSinus(qint32 sampleRate, double delay) :
 }
 
 // Générateur Gordon-Smith
-void OscSinus::getSinus(double * data, qint32 len, double freq)
+void OscSinus::getSinus(float * data, qint32 len, double freq)
 {
     // Attente
     qint64 total = qMin(_delayTime - _currentDelay, len);
